requirements.txt
numpy>=1.21.0
gymnasium>=0.29.0
torch>=2.0.0
Pillow>=9.0.0
matplotlib>=3.5.0
opencv-python>=4.5.0
miniworld>=1.2.0
pyglet>=1.5.0
huggingface-hub>=0.19.0
transformers>=4.35.0
unsloth>=0.3.0
requests>=2.31.0 